+++ main.py
@@ -5,8 +5,9 @@ from sqlalchemy.orm import joinedload
 from sqlalchemy.orm.exc import NoResultFound
 from sqlalchemy.exc import IntegrityError
 from models import User, Chat, ChatStatus
-from debug_routes import debug_routes
+from routes.debug_routes import debug_routes
 from routes.login_route import login_route
+from routes.user_route import user_route
 from db import Session
 import asyncio
 import telebot
@@ -15,6 +16,8 @@ from telebot import types
 from telebot.types import InlineKeyboardMarkup, InlineKeyboardButton, WebAppInfo
 from config import Config
 from shared import user_clients
+from services.user_service import create_user, update_profile
+from services.chat_service import create_chat, add_chat_to_users
 
 commands = (
     "📝 /start - Start the bot\n"
@@ -32,6 +35,7 @@ app = cors(app, allow_origin="*")
 
 app.register_blueprint(debug_routes)
 app.register_blueprint(login_route)
+app.register_blueprint(user_route)
 
 
 async def check_session_expiry():
@@ -76,110 +80,6 @@ async def get_sessions():
         return jsonify(sessions), 200
     except Exception as e:
         return jsonify({"error": str(e)}), 500
-
-async def create_user(user_id, username, profile):
-    session = Session()
-    try:
-        # Query the database to check if a user with the provided ID exists
-        existing_user = session.query(User).filter(User.id == user_id).one()
-        # print("User already exists")
-    except NoResultFound:
-        new_user = User(
-            id=user_id, name=username, has_profile=profile, words=0
-        )
-        user_data = {
-            "id": new_user.id,
-            "name": new_user.name,
-            "has_profile": new_user.has_profile,
-            "words": new_user.words,
-        }
-
-        print(user_data)
-        session.add(new_user)
-        session.commit()
-    except Exception as e:
-        print(f"Error: {str(e)}")
-    finally:
-        session.close()
-
-
-async def create_chat(chat_id, chat_name, words_number, sender_id, chat_users):
-    session = Session()
-    status = 0
-    try:
-        # Query the database to check if a chat with the provided ID exists
-        existing_chat = session.query(Chat).filter(Chat.id == chat_id).one()
-        print("Chat already exists")
-    except NoResultFound:
-        new_chat = Chat(
-            id=chat_id,
-            name=chat_name,
-            words=words_number,
-            status=ChatStatus.pending,
-            lead_id=sender_id,
-            full_text="None",
-        )
-
-        lead = session.query(User).filter(User.id == sender_id).one()
-        new_chat.lead = lead
-
-        agreed_user_ids = [sender_id]
-        agreed_users = session.query(User).filter(User.id.in_(agreed_user_ids)).all()
-        new_chat.agreed_users.extend(agreed_users)
-
-        all_users = (
-            session.query(User).filter(User.id.in_([sender_id] + chat_users)).all()
-        )
-        new_chat.users.extend(all_users)
-
-        session.add(new_chat)
-        session.commit()
-    except Exception as e:
-        print(f"Error: {str(e)}")
-        status = 1
-    finally:
-        session.close()
-        return status
-
-async def update_profile(user_id, has_profile):
-    session = Session()
-    status = 0
-    try:
-        user = session.query(User).filter(User.id == user_id).one()
-        user.has_profile = has_profile
-        session.commit()
-    except NoResultFound:
-        print(f"User with ID {user_id} not found")
-        status = 1
-    except Exception as e:
-        print(f"Error updating username or profile: {str(e)}")
-        status = 1
-    finally:
-        session.close()
-        return status
-
-async def add_chat_to_users(users_id, chat_id):
-    status = 0
-    try:
-        session = Session()
-
-        # get 1 chat
-        chat = session.query(Chat).filter(Chat.id == chat_id).one()
-        # get all related users
-        users = session.query(User).filter(User.id.in_(users_id)).all()
-
-        # add the chat to each user
-        for user in users:
-            if chat not in user.chats:
-                user.chats.append(chat)
-        session.commit()
-    except Exception as e:
-        print(f"Error: {str(e)}")
-        session.rollback()
-        status = 1
-    finally:
-        session.close()
-        return status
 
 
 @app.route("/health", methods=["GET"])
@@ -274,71 +174,6 @@ async def send_message():
             return {"error": str(e)}, 500
 
     return jsonify({"userB": b_users if b_users else None}), 200
-
-
-@app.route("/get-user", methods=["POST"])
-async def get_user():
-    try:
-        data = await request.get_json()
-
-        user_id = data.get("userId")
-        if user_id is None:
-            return jsonify({"error": "userId is missing"}), 400
-        
-        username = data.get("username", "None")
-        print(f"get-user: {username}")
-        try:
-            await create_user(user_id, username, False)
-        except Exception as e:
-            print(f"Error creating user: {str(e)}")
-            return jsonify({"error": "Internal error"}), 500
-        
-        session = Session()
-        
-        # Query all users
-        user = (
-            session.query(User)
-            .options(joinedload(User.chats).joinedload(Chat.users))
-            .filter(User.id == user_id)
-            .first()
-        )
-
-        if user is None:
-            session.close()
-            return jsonify({"message": f"User with id {user_id} does not exist"}), 404
-
-        for chat in user.chats:
-            chat.agreed_users
-
-        # Close the session
-        session.close()
-
-        return jsonify(
-            {
-                "id": user.id,
-                "name": user.name,
-                "has_profile": user.has_profile,
-                "words": user.words,
-                "chats": [
-                    {
-                        "id": chat.id,
-                        "name": chat.name,
-                        "words": chat.words,
-                        "status": chat.status.name,
-                        "lead_id": chat.lead_id,
-                        "agreed_users": [
-                            agreed_user.id for agreed_user in chat.agreed_users
-                        ],
-                        "users": [user.id for user in chat.users],
-                    }
-                    for chat in user.chats
-                ],
-            }
-        )
-
-    except Exception as e:
-        session.close()
-        return jsonify({"error": str(e)}), 500
 
 
 @app.route("/is-active", methods=["POST"])
